Fixed naming issue in shopping cart.
Testing using the browser console, I am able to add items to the shopping cart.
The items display after a refresh, not sure if that's intended or not.

Below is the testing command I used:
sessionStorage.setItem('cart', JSON.stringify([
    { "ProductId" : "54321", "Quantity" : 2},
    { "ProductId" : "12345", "Quantity" : 2}
]));
This sets these items in the cart. ProductId can be changed to match anything in the product list.